refactor(M6 PR2): extract the go-to-file finder into controller/finder.rs
Move the finder feature's methods (open, key + mouse + click handling, confirm,
scroll/selection, the FinderView projection) into a submodule. Cross-dispatch
methods pub(super); intra helpers private. Behavior-preserving. mod.rs ~2737.
